further improve base journey controller
