+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/controllers/rejectedgoodssingle/RejectedGoodsSingleJourneyBaseController.scala
@@ -17,88 +17,22 @@
 package uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.rejectedgoodsingle
 
 import play.api.mvc._
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.cache.SessionCache
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.SessionDataExtractor
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.SessionUpdates
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.actions.AuthenticatedAction
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.actions.SessionDataAction
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.actions.WithAuthAndSessionDataAction
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.journeys.RejectedGoodsSingleJourney
+import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.SessionData
 import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.{upscan => _}
-import uk.gov.hmrc.cdsreimbursementclaimfrontend.utils.Logging
-import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
+import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.JourneyBaseController
 
 import scala.concurrent.ExecutionContext
-import scala.concurrent.Future
 
 abstract class RejectedGoodsSingleJourneyBaseController(
   cc: MessagesControllerComponents
 )(implicit ec: ExecutionContext)
-    extends FrontendController(cc)
-    with WithAuthAndSessionDataAction
-    with SessionDataExtractor
-    with Logging
-    with SessionUpdates {
+    extends JourneyBaseController[RejectedGoodsSingleJourney](cc) {
 
-  val authenticatedAction: AuthenticatedAction
-  val sessionDataAction: SessionDataAction
-  val sessionStore: SessionCache
+  override def getJourney(sessionData: SessionData): Option[RejectedGoodsSingleJourney] =
+    sessionData.rejectedGoodsSingleJourney
 
-  def simpleActionReadJourney(body: RejectedGoodsSingleJourney => Result): Action[AnyContent] =
-    authenticatedActionWithSessionData.async { implicit request =>
-      Future.successful(
-        request.sessionData
-          .flatMap(_.rejectedGoodsSingleJourney)
-          .map(body)
-          .getOrElse(Redirect(uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.routes.StartController.start()))
-      )
-    }
+  override def updateJourney(journey: RejectedGoodsSingleJourney)(sessionData: SessionData): SessionData =
+    sessionData.copy(rejectedGoodsSingleJourney = Some(journey))
 
-  def actionReadJourney(body: Request[_] => RejectedGoodsSingleJourney => Future[Result]): Action[AnyContent] =
-    authenticatedActionWithSessionData.async { implicit request =>
-      request.sessionData
-        .flatMap(_.rejectedGoodsSingleJourney)
-        .map(body(request))
-        .getOrElse(
-          Future.successful(
-            Redirect(uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.routes.StartController.start())
-          )
-        )
-    }
-
-  def simpleActionReadWriteJourney(
-    body: Request[_] => RejectedGoodsSingleJourney => (RejectedGoodsSingleJourney, Result)
-  ): Action[AnyContent] =
-    authenticatedActionWithSessionData.async { implicit request =>
-      request.sessionData
-        .flatMap(_.rejectedGoodsSingleJourney)
-        .map(body(request))
-        .map { case (modifiedJourney, result) =>
-          updateSession(sessionStore, request)(_.copy(rejectedGoodsSingleJourney = Some(modifiedJourney)))
-            .flatMap(_.fold(error => Future.failed(error.toException), _ => Future.successful(result)))
-        }
-        .getOrElse(
-          Future.successful(
-            Redirect(uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.routes.StartController.start())
-          )
-        )
-    }
-
-  def actionReadWriteJourney(
-    body: Request[_] => RejectedGoodsSingleJourney => Future[(RejectedGoodsSingleJourney, Result)]
-  ): Action[AnyContent] =
-    authenticatedActionWithSessionData.async { implicit request =>
-      request.sessionData
-        .flatMap(_.rejectedGoodsSingleJourney)
-        .map(body(request))
-        .map(_.flatMap { case (modifiedJourney, result) =>
-          updateSession(sessionStore, request)(_.copy(rejectedGoodsSingleJourney = Some(modifiedJourney)))
-            .flatMap(_.fold(error => Future.failed(error.toException), _ => Future.successful(result)))
-        })
-        .getOrElse(
-          Future.successful(
-            Redirect(uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.routes.StartController.start())
-          )
-        )
-    }
 }
